# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Talk-to-Learn API
  version: 1.0.0
  description: >-
    OpenAPI-Dokumentation der Server-Endpunkte unter `app/api/*`.
    Die Spezifikation beschreibt die aktuell implementierten Request-/Response-Formate.
servers:
  - url: /
    description: Relative Basis-URL der Next.js App
security:
  - NextAuthSession: []
  - NextAuthSecureSession: []
...
paths:
...
  /api/ai/save:
    post:
      tags: [AI]
      summary: KI-generiertes Lernset speichern
      operationId: saveAiDeck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiSaveRequest'
      responses:
        '200':
          description: Lernset gespeichert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiSaveResponse'
        '400':
          description: Ungültige Anfrage oder planabhängiges Fragenlimit überschritten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/account/dashboard:
    get:
      tags: [Account]
      summary: Account-Dashboard-Daten laden
      operationId: getAccountDashboard
      parameters:
        - in: query
          name: tzOffsetMinutes
          required: false
          schema:
            type: string
          description: Zeitzonenoffset in Minuten zur Berechnung der Tagesnutzung.
      responses:
        '200':
          description: Dashboard geladen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDashboardResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/billing/create-checkout-session:
    post:
      tags: [Billing]
      summary: Stripe Checkout Session erstellen
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutSessionRequest'
      responses:
        '200':
          description: Checkout Session erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCheckoutSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/billing/checkout-status:
    get:
      tags: [Billing]
      summary: Stripe Checkout Session verifizieren und Abo synchronisieren
      operationId: getCheckoutStatus
      parameters:
        - in: query
          name: session_id
          required: true
          schema:
            type: string
          description: Stripe Checkout Session ID (`cs_...`).
      responses:
        '200':
          description: Checkout-Status geladen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Session gehört nicht zum authentifizierten User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    NextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
    NextAuthSecureSession:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
...
  responses:
    BadRequestError:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Nicht authentifiziert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Ressource nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Interner Serverfehler
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AiCard:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
    AiSaveRequest:
      type: object
      required: [title, cards]
      properties:
        title:
          type: string
        cards:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AiCard'
        fileName:
          type: string
    AiSaveResponse:
      type: object
      required: [deckId]
      properties:
        deckId:
          type: string
...
    PlanTier:
      type: string
      enum: [free, premium, ultimate]
    SubscriptionStatus:
      type: string
      enum: [active, cancel_at_period_end, past_due]
    BillingInterval:
      type: [string, 'null']
      enum: [monthly, yearly, null]
    LimitedValue:
      oneOf:
        - type: integer
          minimum: 0
        - type: string
          enum: [unlimited]
    PlanLimits:
      type: object
      required: [deckLimit, questionsPerDeck, speechSecondsPerDay, aiRefine]
      properties:
        deckLimit:
          $ref: '#/components/schemas/LimitedValue'
        questionsPerDeck:
          $ref: '#/components/schemas/LimitedValue'
        speechSecondsPerDay:
          $ref: '#/components/schemas/LimitedValue'
        aiRefine:
          type: boolean
    AccountDashboardResponse:
      type: object
      required:
        - plan
        - status
        - currentPeriodEnd
        - nextBillingAt
        - billingInterval
        - canManageSubscription
        - limits
        - allPlanLimits
        - usage
        - billing
      properties:
        plan:
          $ref: '#/components/schemas/PlanTier'
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        currentPeriodEnd:
          type: [string, 'null']
          format: date-time
        nextBillingAt:
          type: [string, 'null']
          format: date-time
        billingInterval:
          $ref: '#/components/schemas/BillingInterval'
        canManageSubscription:
          type: boolean
        limits:
          $ref: '#/components/schemas/PlanLimits'
        allPlanLimits:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/PlanLimits'
        usage:
          type: object
          required: [speechSecondsToday, decksCreated]
          properties:
            speechSecondsToday:
              type: integer
              minimum: 0
            decksCreated:
              type: integer
              minimum: 0
        billing:
          type: object
          required: [planLabels, planOrder, pricingCards, text]
          properties:
            planLabels:
              type: object
              additionalProperties:
                type: string
            planOrder:
              type: object
              additionalProperties:
                type: integer
            pricingCards:
              type: array
              items:
                type: object
                additionalProperties: true
            text:
              type: object
              additionalProperties: true
    CreateCheckoutSessionRequest:
      type: object
      required: [plan, cycle]
      properties:
        plan:
          type: string
          enum: [premium, ultimate]
        cycle:
          type: string
          enum: [monthly, yearly]
        userEmail:
          type: string
          format: email
        userId:
          type: string
    CreateCheckoutSessionResponse:
      type: object
      required: [url, plan, cycle, configuredPrice]
      properties:
        url:
          type: string
          format: uri
        plan:
          type: string
          enum: [premium, ultimate]
        cycle:
          type: string
          enum: [monthly, yearly]
        configuredPrice:
          type: number
    CheckoutStatusResponse:
      type: object
      required: [checkoutStatus, paymentStatus]
      properties:
        checkoutStatus:
          type: string
        paymentStatus:
          type: [string, 'null']
        synced:
          type: boolean
        subscription:
          $ref: '#/components/schemas/StripeSubscriptionSnapshot'
...
    StripeSubscriptionSnapshot:
      type: object
      required:
        - plan
        - status
        - currentPeriodEnd
        - nextBillingAt
        - billingInterval
        - subscriptionId
        - customerId
        - priceId
      properties:
        plan:
          $ref: '#/components/schemas/PlanTier'
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        currentPeriodEnd:
          type: [string, 'null']
          format: date-time
        nextBillingAt:
          type: [string, 'null']
          format: date-time
        billingInterval:
          $ref: '#/components/schemas/BillingInterval'
        subscriptionId:
          type: [string, 'null']
        customerId:
          type: [string, 'null']
        priceId:
          type: [string, 'null']